requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.0b0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Deprecated==1.2.14
distro==1.9.0
fastapi==0.115.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-core==0.3.12
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langserve==0.3.0
langsmith==0.1.137
multidict==6.1.0
numpy==1.26.4
openai==1.52.2
orjson==3.10.10
packaging==24.1
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sse-starlette==2.1.3
starlette==0.41.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
yarl==1.16.0
